Moved some files and added comments
