+++ src/GIMI-ModManager.Core/GamesService/Models/Element.cs
@@ -1,4 +1,5 @@
 ﻿using System.Diagnostics;
+using GIMI_ModManager.Core.GamesService.Interfaces;
 
 namespace GIMI_ModManager.Core.GamesService.Models;
