+++ src/GIMI-ModManager.Core/GamesService/Models/InternalName.cs
@@ -1,4 +1,5 @@
 ﻿using System.Globalization;
+using GIMI_ModManager.Core.GamesService.Interfaces;
 
 namespace GIMI_ModManager.Core.GamesService.Models;
