+++ src/GIMI-ModManager.WinUI/ViewModels/CharacterGalleryViewModels/SelectableModdableObjectVm.cs
@@ -1,6 +1,5 @@
 ﻿using CommunityToolkit.Mvvm.ComponentModel;
 using CommunityToolkit.Mvvm.Input;
-using GIMI_ModManager.Core.GamesService;
 using GIMI_ModManager.Core.GamesService.Interfaces;
 using GIMI_ModManager.WinUI.Services;
